+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/gnpsexport/fbmn/GNPSExportAndSubmitModule.java
@@ -29,7 +29,7 @@
  * Credit to the Du-Lab development team for the initial commitment to the MGF export module.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.gnpsexport;
+package net.sf.mzmine.modules.peaklistmethods.io.gnpsexport.fbmn;
 
 import java.util.Collection;
 import java.util.logging.Logger;
