+++ src/main/java/net/sf/mzmine/modules/peaklistmethods/io/gnpsexport/fbmn/GNPSExportTask.java
@@ -1,16 +1,16 @@
 /*
  * Copyright 2006-2018 The MZmine 2 Development Team
- *
+ * 
  * This file is part of MZmine 2.
- *
+ * 
  * MZmine 2 is free software; you can redistribute it and/or modify it under the terms of the GNU
  * General Public License as published by the Free Software Foundation; either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
  * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public License along with MZmine 2; if not,
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
@@ -18,27 +18,23 @@
 /*
  * This module was prepared by Abi Sarvepalli, Christopher Jensen, and Zheng Zhang at the Dorrestein
  * Lab (University of California, San Diego).
- *
+ * 
  * It is freely available under the GNU GPL licence of MZmine2.
- *
+ * 
  * For any questions or concerns, please refer to:
  * https://groups.google.com/forum/#!forum/molecular_networking_bug_reports
- *
+ * 
  * Credit to the Du-Lab development team for the initial commitment to the MGF export module.
  */
 
-package net.sf.mzmine.modules.peaklistmethods.io.gnpsexport;
+package net.sf.mzmine.modules.peaklistmethods.io.gnpsexport.fbmn;
 
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.text.DecimalFormat;
-import java.text.MessageFormat;
-import java.text.NumberFormat;
 import java.util.Arrays;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import java.util.regex.Pattern;
+
 import net.sf.mzmine.datamodel.DataPoint;
 import net.sf.mzmine.datamodel.Feature;
 import net.sf.mzmine.datamodel.MassList;
@@ -48,7 +44,6 @@ import net.sf.mzmine.datamodel.Scan;
 import net.sf.mzmine.datamodel.impl.SimpleFeature;
 import net.sf.mzmine.datamodel.impl.SimplePeakListRow;
 import net.sf.mzmine.main.MZmineCore;
-import net.sf.mzmine.modules.peaklistmethods.io.gnpsexport.GNPSExportAndSubmitParameters.RowFilter;
 import net.sf.mzmine.modules.tools.msmsspectramerge.MergedSpectrum;
 import net.sf.mzmine.modules.tools.msmsspectramerge.MsMsSpectraMergeModule;
 import net.sf.mzmine.modules.tools.msmsspectramerge.MsMsSpectraMergeParameters;
@@ -56,46 +51,32 @@ import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.taskcontrol.AbstractTask;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 import net.sf.mzmine.util.PeakUtils;
-import net.sf.mzmine.util.files.FileAndPathUtil;
 
 /**
  * Exports all files needed for GNPS
- *
+ * 
  * @author Robin Schmid (robinschmid@uni-muenster.de)
  *
  */
-public class GNPSmgfExportTask extends AbstractTask {
-  // Logger.
-  private final Logger LOG = Logger.getLogger(getClass().getName());
-
-  //
+public class GNPSExportTask extends AbstractTask {
   private final PeakList[] peakLists;
   private final File fileName;
   private final String plNamePattern = "{}";
   private int currentIndex = 0;
-  private final String massListName;
   private final MsMsSpectraMergeParameters mergeParameters;
+  
+  private final String massListName;
 
-  // by robin
-  private NumberFormat mzForm = MZmineCore.getConfiguration().getMZFormat();
-  private NumberFormat intensityForm = MZmineCore.getConfiguration().getIntensityFormat();
-  // seconds
-  private NumberFormat rtsForm = new DecimalFormat("0.###");
-  // correlation
-  private NumberFormat corrForm = new DecimalFormat("0.0000");
-
-  private RowFilter filter;
-
-  GNPSmgfExportTask(ParameterSet parameters) {
-    this.peakLists = parameters.getParameter(GNPSExportAndSubmitParameters.PEAK_LISTS).getValue()
-        .getMatchingPeakLists();
+  GNPSExportTask(ParameterSet parameters) {
+    this.peakLists =
+        parameters.getParameter(GNPSExportAndSubmitParameters.PEAK_LISTS).getValue().getMatchingPeakLists();
 
     this.fileName = parameters.getParameter(GNPSExportAndSubmitParameters.FILENAME).getValue();
+
     this.massListName = parameters.getParameter(GNPSExportAndSubmitParameters.MASS_LIST).getValue();
-    this.filter = parameters.getParameter(GNPSExportAndSubmitParameters.FILTER).getValue();
+
     if (parameters.getParameter(GNPSExportAndSubmitParameters.MERGE_PARAMETER).getValue()) {
-      mergeParameters = parameters.getParameter(GNPSExportAndSubmitParameters.MERGE_PARAMETER)
-          .getEmbeddedParameters();
+      mergeParameters = parameters.getParameter(GNPSExportAndSubmitParameters.MERGE_PARAMETER).getEmbeddedParameters();
     } else {
       mergeParameters = null;
     }
@@ -130,7 +111,6 @@ public class GNPSmgfExportTask extends AbstractTask {
             fileName.getPath().replaceAll(Pattern.quote(plNamePattern), cleanPlName);
         curFile = new File(newFilename);
       }
-      curFile = FileAndPathUtil.getRealFilePath(curFile, "mgf");
 
       // Open file
       FileWriter writer;
@@ -174,20 +154,13 @@ public class GNPSmgfExportTask extends AbstractTask {
       setStatus(TaskStatus.FINISHED);
   }
 
-  private int export(PeakList peakList, FileWriter writer, File curFile) throws IOException {
+  private void export(PeakList peakList, FileWriter writer, File curFile) throws IOException {
     final String newLine = System.lineSeparator();
 
-    // count exported
-    int count = 0;
-    int countMissingMassList = 0;
     for (PeakListRow row : peakList.getRows()) {
-      // do not export if no MSMS
-      if (!filter.filter(row))
-        continue;
-
       String rowID = Integer.toString(row.getID());
-      double retTimeInSeconds = ((row.getAverageRT() * 60 * 100.0) / 100.);
 
+      String retTimeInSeconds = Double.toString(Math.round(row.getAverageRT() * 60 * 100) / 100.);
       // Get the MS/MS scan number
       Feature bestPeak = row.getBestPeak();
       if (bestPeak == null)
@@ -198,15 +171,10 @@ public class GNPSmgfExportTask extends AbstractTask {
         // Best peak always exists, because feature list row has at least one peak
         bestPeak = copyRow.getBestPeak();
 
-        // Get the heighest peak with a MS/MS scan number (with mass list)
-        boolean missingMassList = false;
-        msmsScanNumber = bestPeak.getMostIntenseFragmentScanNumber();
-        while (msmsScanNumber < 1
-            || getScan(bestPeak, msmsScanNumber).getMassList(massListName) == null) {
-          // missing masslist
-          if (msmsScanNumber > 0)
-            missingMassList = true;
+        // Get the MS/MS scan number
 
+        msmsScanNumber = bestPeak.getMostIntenseFragmentScanNumber();
+        while (msmsScanNumber < 1) {
           copyRow.removePeak(bestPeak.getDataFile());
           if (copyRow.getPeaks().length == 0)
             break;
@@ -214,8 +182,6 @@ public class GNPSmgfExportTask extends AbstractTask {
           bestPeak = copyRow.getBestPeak();
           msmsScanNumber = bestPeak.getMostIntenseFragmentScanNumber();
         }
-        if (missingMassList)
-          countMissingMassList++;
       }
       if (msmsScanNumber >= 1) {
         // MS/MS scan must exist, because msmsScanNumber was > 0
@@ -223,9 +189,11 @@ public class GNPSmgfExportTask extends AbstractTask {
 
         MassList massList = msmsScan.getMassList(massListName);
 
-
         if (massList == null) {
-          continue;
+          MZmineCore.getDesktop().displayErrorMessage(MZmineCore.getDesktop().getMainWindow(),
+                  "There is no mass list called " + massListName + " for MS/MS scan #" + msmsScanNumber
+                          + " (" + bestPeak.getDataFile() + ")");
+          return;
         }
 
         writer.write("BEGIN IONS" + newLine);
@@ -233,13 +201,13 @@ public class GNPSmgfExportTask extends AbstractTask {
         if (rowID != null)
           writer.write("FEATURE_ID=" + rowID + newLine);
 
-        String mass = mzForm.format(row.getAverageMZ());
+        String mass = Double.toString(Math.round(row.getAverageMZ() * 10000) / 10000.);
         if (mass != null)
           writer.write("PEPMASS=" + mass + newLine);
 
         if (rowID != null) {
           writer.write("SCANS=" + rowID + newLine);
-          writer.write("RTINSECONDS=" + rtsForm.format(retTimeInSeconds) + newLine);
+          writer.write("RTINSECONDS=" + retTimeInSeconds + newLine);
         }
 
         int msmsCharge = msmsScan.getPrecursorCharge();
@@ -253,14 +221,11 @@ public class GNPSmgfExportTask extends AbstractTask {
         writer.write("CHARGE=" + msmsCharge + msmsPolarity + newLine);
 
         writer.write("MSLEVEL=2" + newLine);
-
         DataPoint[] dataPoints = massList.getDataPoints();
         if (mergeParameters != null) {
-          MsMsSpectraMergeModule merger =
-              MZmineCore.getModuleInstance(MsMsSpectraMergeModule.class);
-          MergedSpectrum spectrum =
-              merger.getBestMergedSpectrum(mergeParameters, row, massListName);
-          if (spectrum != null) {
+          MsMsSpectraMergeModule merger = MZmineCore.getModuleInstance(MsMsSpectraMergeModule.class);
+          MergedSpectrum spectrum = merger.getBestMergedSpectrum(mergeParameters, row, massListName);
+          if (spectrum!=null) {
             dataPoints = spectrum.data;
             writer.write("MERGED_STATS=");
             writer.write(spectrum.getMergeStatsDescription());
@@ -268,35 +233,17 @@ public class GNPSmgfExportTask extends AbstractTask {
           }
         }
         for (DataPoint peak : dataPoints) {
-          writer.write(mzForm.format(peak.getMZ()) + " " + intensityForm.format(peak.getIntensity())
-              + newLine);
+          writer.write(peak.getMZ() + " " + peak.getIntensity()
+                  + newLine);
         }
+
         writer.write("END IONS" + newLine);
         writer.write(newLine);
-        count++;
       }
     }
 
-    if (count == 0)
-      LOG.log(Level.WARNING, "No MS/MS scans exported.");
-    else
-      LOG.info(
-          MessageFormat.format("Total of {0} feature rows (MS/MS mass lists) were exported ({1})",
-              count, peakList.getName()));
-
-    if (countMissingMassList > 0)
-      LOG.warning(MessageFormat.format(
-          "WARNING: Total of {0} feature rows have an MS/MS scan but NO mass list (this shouldn't be a problem if a scan filter was applied in the mass detection step) ({1})",
-          countMissingMassList, peakList.getName()));
-
-    return count;
   }
 
-  public Scan getScan(Feature f, int msmsscan) {
-    return f.getDataFile().getScan(msmsscan);
-  }
-
-  @Override
   public String getTaskDescription() {
     return "Exporting GNPS of feature list(s) " + Arrays.toString(peakLists) + " to MGF file(s)";
   }
